obradjene greske za istaknute blogove

+++ app/Http/Controllers/BlogController.php
@@ -47,14 +47,23 @@ class BlogController extends Controller
         return $blogovi;
     }
 
+    public static function broj_istaknutih()
+    {
+        $broj_istaknutih= Blog::where('istaknut', 1)->count();
+
+        return $broj_istaknutih;
+    }
+
     //crud metode
     public function list()
     {
+        $broj_istaknutih= Blog::where('istaknut', 1)->count();
         $blogovi = Blog::all();
 
         return view('blog.list', [
             'blogovi' => $blogovi,
             'title' => 'Листа блогова',
+            'broj_istaknutih'=> $broj_istaknutih
         ]);
     }
 
@@ -150,6 +159,9 @@ class BlogController extends Controller
     public function unpublish($id)
     {
         $blog = Blog::find($id);
+        if ($blog->istaknut==1) {
+            return redirect(route('blog.list'))->with('info', 'Блог не може бити сакривен ако је истакнут.');
+        }
         if (!$blog) {
             return abort(404);
         }
@@ -162,13 +174,21 @@ class BlogController extends Controller
     public function istakni($id)
     {
         $blog = Blog::find($id);
-        if (!$blog) {
-            return abort(404);
+        if ($blog->objavljen==0) {
+            return redirect(route('blog.list'))->with('info', 'Блог не може бити истакнут ако је сакривен.');
         }
-        $blog->istaknut = 1;
-        $blog->save();
-
-        return redirect(route('blog.list'));
+        $broj_istaknutih= Blog::where('istaknut', 1)->count();
+        if ($broj_istaknutih<7) {
+            if (!$blog) {
+                return abort(404);
+            }
+            $blog->istaknut = 1;
+            $blog->save();
+            return redirect(route('blog.list'));
+        }
+        else {
+            return redirect(route('blog.list'))->with('info', 'Максималан број истакнутих блогова је 7.');
+        }
     }
 
     public function obrisi_istakni($id)
@@ -188,6 +208,9 @@ class BlogController extends Controller
         $blog = Blog::find($id);
         if (!$blog) {
             return abort(404);
+        }
+        if ($blog->istaknut=1) {
+            return redirect(route('blog.list'))->with('info', 'Блогови који су истакнути не могу бити обрисани.');
         }
         $blog->delete();
